requirements.txt
matplotlib==3.3.4
numpy==1.20.2
dash
requests
flask
retrying
ipython
ipykernel
ansi2html
dash_html_components==1.1.4
pandas==1.2.4
dash_core_components==1.17.1
plotly>=5.2.1
jupyterlab>=2.0
notebook>=6.0
jupyterlab_launcher>=0.5.0
jupyter_server_proxy
jupyterlab-dash
jupyter-dash
pycoingecko
dash_bootstrap_components
dash_html_components
dash_table
streamlit==0.89.0
toml
pandas
numpy
matplotlib
plotly==5.3.1
plotly_express==0.4.0
pycoingecko
DateTime
millify
streamlit-lottie>=0.0.3
Pillow~=8.4.0